Refactor PinotMetricUtils and MetricsRegistryRegistrationListener related classes from pinot-common to pinot-spi
