+++ pinot-spi/src/main/java/org/apache/pinot/spi/metrics/MetricsRegistryRegistrationListener.java
@@ -16,9 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.metrics;
+package org.apache.pinot.spi.metrics;
 
-import org.apache.pinot.spi.metrics.PinotMetricsRegistry;
 
 
 /**
